python/requirements.txt
aiohttp~=3.7.3
attrs~=19.3.0
numpy~=1.19.5
setuptools~=45.2.0
networkx~=2.5
python-igraph~=0.8.3
ray~=1.1.0
psutil~=5.8.0
codetiming~=1.2.0
jsonpickle~=1.5.0
boto3~=1.16.57